Update patch set 1
Patch Set 1:

This is what Samsung uses.

It is, at least in theory, more appropriate to Exynos.  One thing not included is a cpu.awake value - we don't know what the consumption is if a wakelock is held but the CPU is inactive.

The question is - do we take a guess at the cpu.awake value, or ignore it for now?  It's probably on the order of 10-20, just judging by the fact that when wakelocked and cpuidle is working, drain is 1.25-1.5%/hour, compared to around 4-5%/hour if cpuidle is busted.

Patch-set: 1