Update patch set 1
Patch Set 1: No score

thanks, I didn't notice that commit, building with this and mentioned commit and will update later if I have some more info.

I did a quick google on the RIL crashing issue and it seems that setting minimal cpu to 500mhz might help as well, not sure if it is related, just thought I mention it.

Patch-set: 1