Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Indeed, the libaudio-ril-interface stuff needs to be changed to make use of libsecril-client. That's probably not too hard. 

Also, there is no support for WB AMR in TinyALSA-Audio, but AFAIK, it's a parameter set via set_parameters (on audio_hw or audio_out, can't recall), then it is forwarded to the RIL client (which forwards it to the RIL).

Patch-set: 1