Update patch set 4
Patch Set 4:

Hmm, only applies to 4210 devices, and maybe not even all of them.

There's a lot of galaxys2 customization - of course the tabs might have nearly identical routing/mixer settings.  You might need to add a new file to yamaha-mc1n2-audio/devices/

As to bluetooth SCO - it's not quite working.  It seems like the modem mixer settings in the XML are only applied when the modem starts - if you're incall, but it's a mixer settings change, the output settings are applied instead.

Patch-set: 4