Update patch set 5
Patch Set 5:

daniel: cool you thought about this, thanks! however, iirc, ive debugged to the point of uusPresent==1, and the memset in that would fail. the memset you added rhere is for the uusPresent==0 case. i might be remembering it wrong, and if you aint sure as well, id be giving uusPresent==1 a shot... ;)

Patch-set: 5