Update patch set 1
Patch Set 1:

I did not check CM gerrit until two weeks when I sent n7000 patches, so it is why I did not change authorship, because when I did there was no one sent :)

And that does not fix any troubles, is already tested and Andrew gave him idea about it, I just suggested. If CM wanna merge go ahead with it, but I do not think it is needed.

Patch-set: 1