Update patch set 1
Patch Set 1:

Removing our custom releasetools completely breaks the boot.img in the OTA zip, so that's the only piece that we still need different than CM releasetools.
Initially I thought about updating the backuptools only which appears to be broken in the first nightly, but then all of it is bad, it takes about 5 minutes to update with the old releasetools, while CM's releasetools takes a few seconds only.

Patch-set: 1