Update patch set 5
Patch Set 5:

hi Arnab,

i have doubts regarding the choice of F2FS mount options, i think we should review them.

the options are documented here:
https://www.kernel.org/doc/Documentation/filesystems/f2fs.txt


1) discard

f2fs, as merged into the i9100 kernel, does not support batch discard (ie, fstrim):
http://forum.xda-developers.com/galaxy-s2/development-derivatives/rom-brickbug-aftermath-speeding-t2843238/post65569217

according to old posts, at the time this wasn't implemented:
https://sourceforge.net/p/linux-f2fs/mailman/message/32515474/

however it looks like maybe newer f2fs versions do support it and the version merged into i9100 is just old:
http://forum.xda-developers.com/showpost.php?p=65569545&postcount=1123

so...

the best course would be to upgrade f2fs to a more capable version.

but...

as stopgap measure i strongly recommend discard be added to mount options. right now there is NO WAY to trim and that is REALLY BAD. by adding discard, at least those with trim enabled kernels will be able to have working -but slow- online trim. otherwise f2fs is really not usable on the platform IMHO.

the discard option will not affect stock kernel users (but will give users the choice to enable trim by replacing the kernel).

note that the discard option should be removed if the f2fs implementation is ever upgraded to support batch discard.


2) active_logs

the default is 6, but this has been changed to 2. unless there is strong justification for the change, i suggest we erase this option. the file system designers chose 6, and i trust them much more than amateur hackers (all of us i think) that like to fiddle with stuff they don't fully understand.


3) background_gc

same thing: defaults to on but has been inexplicably turned off. no background gc means GC is done inline with urgent operations, which is SLOOOOWWWWW. we should remove this option unless there is strong justification for turning it off.

also, inline discard while on the background should not be that bad. so even if we can't upgrade to batch discard, enabling this option could help us live with the current limitation.


4) inline_xattr, inline_data, inline_dentry

these options increase performance in a safe manner. AFAIK, these exist as options only to maintain partition compatibility with older f2fs implementations. we don't care about compatibility with old kernels, so all of these should be enabled IMHO.

note that /cache and /data currently have different settings, without explanation of any kind.


5) flush_merge

on the i9100, our CPUs are way faster than our eMMCs. so without proper profiling tests, my guesstimate is that this option should be enabled.


6) data_flush

enabling this option will slow things down. but it might also make buggy tools (of which there are zillions in the unix world) behave safely. you can read more about how ext4 had to be hacked -although ext4 wasn't the problem- to support all these buggy programs floating around:
https://lwn.net/Articles/322823/

f2fs is believed to be correct, but it does not incorporate these ext4-style hacks yet AFAIK (can't find the citation now but i did read about this). this means that buggy software which was nonetheless safe on ext4 becomes unsafe on f2fs in case of kernel crashes or unclean powerdowns.

i suspect -but i'm not sure- that this option might make f2fs as safe as ext4 in this regard, of course trading off some performance for safety.

i recommend that this be researched further and this option added if warranted. or even better: add it right now just to be on the safe side, and disable it later if justified. we have a lot of freezes and reboots on this platform still and for now we have to be extremely cautious in order to avoid corruption, data loss, bootloops etc.


thanks!

Patch-set: 5
Reviewer: Lanchon <9776@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0