Update patch set 1
Patch Set 1: -Code-Review

i understand.

with a kernel that crashes i think it's dangerous not to enable data_flush, without prior knowledge that data loss won't happen during crashes. i can't +1 this.

but given that the risk is already committed to the tree, i have no reason to oppose this either (other than we could do better), so i'll remove the -1.

regarding flush_merge, it's a guesstimate as explained in my comment here:
http://review.cyanogenmod.org/#/c/132900/5
if you believe you have reasons to favor not including it, they are a priori just as valid. i don't plan to test or benchmark anything myself. but maybe my argument sounded convincing given that we are choosing blindly here.

regarding the difference in options among cache and data, i can't see a rationale for that.

thanks!

Patch-set: 1