Update patch set 1
Patch Set 1:

Not in a hurry, didn't ask for any ETA-s or ASAP-s.

Decided to comment because you said "without benchmarks I won't add data_flush", which I understand as "without proof it doesn't slow performance I will not be merging changes that improve stability", which to me is the wrong attitude WRT data integrity.

No offence meant, sorry if you felt it was.

Patch-set: 1