Update patch set 1
Patch Set 1:

Exact same thing which I was trying to tell you before in our conversation. Replace the INSTALLED_KERNEL_TARGET dependency with recovery_uncompressed_ramdisk and then modify the tool used in the process in ln 34. Using the same with INSTALLED_KERNEL_TARGET dependency on ln 33 and running the lzo command was dirty.

Patch-set: 1