Update patch set 1
Patch Set 1:

(1 comment)

Somehow the ls-command (line 52) is returning an error-code because of access denied on .android_secure-folder but still printing the directory contents. In M it doesn't return an error-code regardless of the access denied on .android_secure-folder. My guess maybe our kernel is missing something or it was a bug in M and now fixed in N causing CMFileManager not to access the storages containing .android_secure. In the command_list.xml in line 52 I changed && into ; so CMFileManager parses the result.

Patch-set: 1